Report the runner's host specs to the control plane on ping
The control plane had no idea what hardware a runner was on. The Runner
model records architecture and region but nothing about size, so the
dashboard could not answer "what does my runner actually have", and
there was no value to validate a future resource setting against.

The runner now sends host memory, vCPU count and EC2 instance type with
its ping. It has to be the runner rather than the aws-controller: the
controller is a Fargate task with no view of the EC2 host. Because the
runner scales to zero between jobs these ride the ping it sends when it
boots, so the stored values describe the last boot -- correct until the
instance is resized.

Host detection moves to utils/hostinfo. It cannot stay in jobs/commands
because the ping client needs it too and jobs/commands already imports
client, which would make that a cycle. jobs/commands keeps thin
delegating wrappers so the sizing code reads unchanged.

Instance type comes from IMDS, bounded by a short timeout and cached: it
is a display nicety and must never delay the first ping, which is what
marks a runner alive. It returns empty off EC2. Note this is the RUNNER
reaching IMDS, which is expected -- it is the agentbox and build
containers that are firewalled off from 169.254.169.254 so untrusted
code cannot reach the host's IAM credentials.

All three values are best-effort and may be zero/empty, and the model
fields will be omitempty so a ping without them never clobbers a
previously reported value.
